+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Fix `<al-field-schema>#getSchemaSupplierField()` not finding the supplier field when it's a nested schema
 * Don't let `<al-field-schema>#sub_fields` return meta fields
 * Fix `<al-field-schema>` having issues with nested `<al-field-schema>` fields
+* Make `<al-field>` remember forced `schema` values
 
 ## 0.3.0-alpha.2 (2024-02-25)